refactor: extract path resolution into homeboy-paths crate
Moves the paths module (config/data/artifact/rig/runtime dir resolution,
remote-path authorization, ~900 LOC across paths.rs + locations/rigs/runtime)
into a new path-only crate. Depends only on the already-extracted
homeboy-error and homeboy-product-identity crates plus shellexpand — made
possible by the earlier paths<->defaults cycle break, which removed paths'
runtime dependency on the config layer.

A few path utilities used across the monolith (sanitize_path_segment,
join_remote_child, resolve_optional_base_path, runner_sessions_dir) are
promoted from pub(crate) to pub so they remain reachable across the new crate
boundary.

The paths test module depended on test_support (HOME isolation) and the
defaults config layer, neither of which belong in a leaf crate, so it was
relocated to src/core/paths_tests.rs and now exercises the crate through
crate::core::paths. The pure rig-path tests stay in the crate. The config
override precedence tests (from the cycle break) are preserved.

Re-exported via core/mod.rs so crate::core::paths::* call sites are unchanged.
Not published to crates.io; still ships as the single homeboy binary.
cargo check --lib and cargo check --lib --tests both pass.
