+++ Cargo.toml
@@ -40,6 +40,7 @@ path = "src/bin/bench-audit-self.rs"
 homeboy-error = { path = "crates/homeboy-error" }
 homeboy-finding = { path = "crates/homeboy-finding" }
 homeboy-output = { path = "crates/homeboy-output" }
+homeboy-paths = { path = "crates/homeboy-paths" }
 homeboy-process = { path = "crates/homeboy-process" }
 homeboy-product-identity = { path = "crates/homeboy-product-identity" }
 homeboy-engine-primitives = { path = "crates/homeboy-engine-primitives" }
