+++ crates/homeboy-paths/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "homeboy-paths"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+authors = ["Chris Huber <chubes@extrachill.com>"]
+description = "Filesystem path resolution (config/data/artifact/rig/runtime dirs) for homeboy"
+repository = "https://github.com/Extra-Chill/homeboy"
+
+[lib]
+name = "homeboy_paths"
+path = "src/lib.rs"
+
+[dependencies]
+homeboy-error = { path = "../homeboy-error" }
+homeboy-product-identity = { path = "../homeboy-product-identity" }
+shellexpand = "3.1"
