+++ crates/homeboy-paths/src/lib.rs
@@ -1,4 +1,4 @@
-use crate::core::error::{Error, Result};
+use homeboy_error::{Error, Result};
 use std::env;
 use std::ffi::OsString;
 use std::path::{Component, Path, PathBuf};
@@ -72,8 +72,7 @@ pub fn homeboy() -> Result<PathBuf> {
                 "APPDATA environment variable not set on Windows".to_string(),
             )
         })?;
-        Ok(PathBuf::from(appdata)
-            .join(crate::core::product_identity::PRODUCT_IDENTITY.config_dirname))
+        Ok(PathBuf::from(appdata).join(homeboy_product_identity::PRODUCT_IDENTITY.config_dirname))
     }
 
     #[cfg(not(windows))]
@@ -85,13 +84,13 @@ pub fn homeboy() -> Result<PathBuf> {
         })?;
         Ok(PathBuf::from(home)
             .join(".config")
-            .join(crate::core::product_identity::PRODUCT_IDENTITY.config_dirname))
+            .join(homeboy_product_identity::PRODUCT_IDENTITY.config_dirname))
     }
 }
 
 /// Global product config file path
 pub fn homeboy_json() -> Result<PathBuf> {
-    Ok(crate::core::product_identity::PRODUCT_IDENTITY.config_file(homeboy()?))
+    Ok(homeboy_product_identity::PRODUCT_IDENTITY.config_file(homeboy()?))
 }
 
 /// Base Homeboy data directory for local observed state.
@@ -108,7 +107,7 @@ pub fn homeboy_data() -> Result<PathBuf> {
                     "LOCALAPPDATA or APPDATA environment variable not set on Windows".to_string(),
                 )
             })?;
-        Ok(PathBuf::from(base).join(crate::core::product_identity::PRODUCT_IDENTITY.data_dirname))
+        Ok(PathBuf::from(base).join(homeboy_product_identity::PRODUCT_IDENTITY.data_dirname))
     }
 
     #[cfg(not(windows))]
@@ -116,7 +115,7 @@ pub fn homeboy_data() -> Result<PathBuf> {
         if let Ok(xdg_data_home) = env::var("XDG_DATA_HOME") {
             if !xdg_data_home.trim().is_empty() {
                 return Ok(PathBuf::from(xdg_data_home)
-                    .join(crate::core::product_identity::PRODUCT_IDENTITY.data_dirname));
+                    .join(homeboy_product_identity::PRODUCT_IDENTITY.data_dirname));
             }
         }
 
@@ -128,7 +127,7 @@ pub fn homeboy_data() -> Result<PathBuf> {
         Ok(PathBuf::from(home)
             .join(".local")
             .join("share")
-            .join(crate::core::product_identity::PRODUCT_IDENTITY.data_dirname))
+            .join(homeboy_product_identity::PRODUCT_IDENTITY.data_dirname))
     }
 }
 
@@ -174,7 +173,7 @@ pub fn expand_tilde_path(path: impl AsRef<Path>) -> PathBuf {
     PathBuf::from(shellexpand::tilde(&raw).into_owned())
 }
 
-pub(crate) fn sanitize_path_segment(segment: &str) -> String {
+pub fn sanitize_path_segment(segment: &str) -> String {
     segment
         .chars()
         .map(|ch| {
@@ -328,7 +327,7 @@ pub fn resolve_path_string(base: &str, file: &str) -> String {
     resolve_path(base, file).to_string_lossy().to_string()
 }
 
-pub(crate) fn resolve_optional_base_path(base_path: Option<&str>) -> Option<&str> {
+pub fn resolve_optional_base_path(base_path: Option<&str>) -> Option<&str> {
     base_path.and_then(|value| (!value.trim().is_empty()).then_some(value.trim()))
 }
 
@@ -359,7 +358,7 @@ pub fn join_remote_path(base_path: Option<&str>, path: &str) -> Result<String> {
     }
 }
 
-pub(crate) fn join_remote_child(base_path: Option<&str>, dir: &str, child: &str) -> Result<String> {
+pub fn join_remote_child(base_path: Option<&str>, dir: &str, child: &str) -> Result<String> {
     let dir_path = join_remote_path(base_path, dir)?;
     let child = child.trim();
 
@@ -438,235 +437,5 @@ pub fn authorize_remote_artifact_path(
         Ok(())
     } else {
         Err(RemotePathAuthorizationError::OutsideAllowedRoots)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_support::with_isolated_home;
-
-    #[test]
-    fn join_remote_path_allows_absolute_paths_without_base() {
-        assert_eq!(
-            join_remote_path(None, "/var/log/syslog").unwrap(),
-            "/var/log/syslog"
-        );
-    }
-
-    #[test]
-    fn artifact_root_defaults_under_homeboy_data() {
-        with_isolated_home(|home| {
-            assert_eq!(
-                artifact_root().expect("artifact root"),
-                home.path().join(".local/share/homeboy/artifacts")
-            );
-        });
-    }
-
-    #[test]
-    fn expand_tilde_path_uses_home_and_preserves_other_paths() {
-        with_isolated_home(|home| {
-            assert_eq!(expand_tilde_path("~/source"), home.path().join("source"));
-            assert_eq!(
-                expand_tilde_path("relative/source"),
-                PathBuf::from("relative/source")
-            );
-        });
-    }
-
-    #[test]
-    fn artifact_root_uses_configured_value() {
-        with_isolated_home(|home| {
-            // Mirror production startup wiring: register the config resolver so
-            // artifact_root() can read the config-level override.
-            set_config_artifact_root_resolver(|| {
-                crate::core::defaults::load_config().artifact_root
-            });
-            let configured = home.path().join("custom-artifacts");
-            crate::core::defaults::save_config(&crate::core::defaults::HomeboyConfig {
-                artifact_root: Some(configured.to_string_lossy().to_string()),
-                ..crate::core::defaults::HomeboyConfig::default()
-            })
-            .expect("save config");
-
-            assert_eq!(artifact_root().expect("artifact root"), configured);
-        });
-    }
-
-    #[test]
-    fn artifact_root_prefers_env_over_config() {
-        with_isolated_home(|home| {
-            set_config_artifact_root_resolver(|| {
-                crate::core::defaults::load_config().artifact_root
-            });
-            let configured = home.path().join("config-artifacts");
-            let env_root = home.path().join("env-artifacts");
-            crate::core::defaults::save_config(&crate::core::defaults::HomeboyConfig {
-                artifact_root: Some(configured.to_string_lossy().to_string()),
-                ..crate::core::defaults::HomeboyConfig::default()
-            })
-            .expect("save config");
-            std::env::set_var("HOMEBOY_ARTIFACT_ROOT", &env_root);
-
-            assert_eq!(artifact_root().expect("artifact root"), env_root);
-        });
-    }
-
-    #[test]
-    fn test_set_artifact_root_override() {
-        with_isolated_home(|home| {
-            let env_root = home.path().join("env-artifacts");
-            let override_root = home.path().join("override-artifacts");
-            std::env::set_var("HOMEBOY_ARTIFACT_ROOT", &env_root);
-            set_artifact_root_override(Some(override_root.clone()));
-
-            assert_eq!(artifact_root().expect("artifact root"), override_root);
-        });
-    }
-
-    #[test]
-    fn join_remote_path_rejects_relative_paths_without_base() {
-        assert!(join_remote_path(None, "file.json").is_err());
-    }
-
-    #[test]
-    fn join_remote_path_joins_relative_paths() {
-        assert_eq!(
-            join_remote_path(Some("/var/www/site"), "file.json").unwrap(),
-            "/var/www/site/file.json"
-        );
-
-        assert_eq!(
-            join_remote_path(Some("/var/www/site/"), "file.json").unwrap(),
-            "/var/www/site/file.json"
-        );
-    }
-
-    #[test]
-    fn join_remote_child_appends_child() {
-        assert_eq!(
-            join_remote_child(Some("/var/www/site"), "logs", "error.log").unwrap(),
-            "/var/www/site/logs/error.log"
-        );
-
-        assert_eq!(
-            join_remote_child(Some("/var/www/site"), "/var/log", "syslog").unwrap(),
-            "/var/log/syslog"
-        );
-    }
-
-    #[test]
-    fn authorize_remote_artifact_path_checks_lexical_policy() {
-        let roots = vec!["/runner/workspace/".to_string()];
-        assert_eq!(
-            authorize_remote_artifact_path(
-                Path::new("relative"),
-                &roots,
-                RemotePathRootContainment::RemoteString,
-            ),
-            Err(RemotePathAuthorizationError::NotAbsolute)
-        );
-        assert_eq!(
-            authorize_remote_artifact_path(
-                Path::new("/runner/workspace/../secret"),
-                &roots,
-                RemotePathRootContainment::RemoteString,
-            ),
-            Err(RemotePathAuthorizationError::ContainsParentDir)
-        );
-        assert_eq!(
-            authorize_remote_artifact_path(
-                Path::new("/other/artifact"),
-                &roots,
-                RemotePathRootContainment::RemoteString,
-            ),
-            Err(RemotePathAuthorizationError::OutsideAllowedRoots)
-        );
-        assert!(authorize_remote_artifact_path(
-            Path::new("/runner/workspace/out"),
-            &roots,
-            RemotePathRootContainment::RemoteString,
-        )
-        .is_ok());
-    }
-
-    #[test]
-    fn resolve_optional_base_path_trims_and_rejects_empty() {
-        assert_eq!(
-            resolve_optional_base_path(Some(" /var/www ")),
-            Some("/var/www")
-        );
-        assert_eq!(resolve_optional_base_path(Some("   ")), None);
-        assert_eq!(resolve_optional_base_path(None), None);
-    }
-
-    #[test]
-    fn normalize_local_path_collapses_dot_and_parent_segments() {
-        assert_eq!(
-            normalize_local_path("/repo/./packages/../src"),
-            PathBuf::from("/repo/src")
-        );
-        assert_eq!(
-            normalize_local_path("packages/../src/./lib"),
-            PathBuf::from("src/lib")
-        );
-        assert_eq!(
-            normalize_local_path("../../src"),
-            PathBuf::from("../../src")
-        );
-        assert_eq!(normalize_local_path("/../../src"), PathBuf::from("/src"));
-    }
-
-    #[test]
-    fn local_path_containment_is_component_aware() {
-        assert!(local_path_is_contained("/repo", "/repo/src/lib.rs"));
-        assert!(local_path_is_contained("/repo", "/repo/src/../Cargo.toml"));
-        assert!(!local_path_is_contained("/repo", "/repo-other/file"));
-        assert!(!local_path_is_contained("/repo", "/repo/../etc/passwd"));
-    }
-
-    #[test]
-    fn resolve_contained_local_path_resolves_relative_paths_under_root() {
-        assert_eq!(
-            resolve_contained_local_path("/repo", "src/../Cargo.toml", "cwd").unwrap(),
-            PathBuf::from("/repo/Cargo.toml")
-        );
-    }
-
-    #[test]
-    fn resolve_contained_local_path_rejects_parent_escape() {
-        let err = resolve_contained_local_path("/repo/worktree", "../secrets", "cwd")
-            .expect_err("parent escape should fail");
-
-        assert!(err.to_string().contains("escapes root '/repo/worktree'"));
-    }
-
-    #[test]
-    fn resolve_path_handles_relative() {
-        let result = resolve_path_string("/base", "relative/path");
-        assert_eq!(result, "/base/relative/path");
-    }
-
-    #[test]
-    fn test_runner_sessions_dir_under_homeboy_dir() {
-        let path = runner_sessions_dir().expect("runner_sessions_dir resolves");
-        assert!(path.ends_with("runner-sessions"), "got {}", path.display());
-        assert!(path.parent().expect("parent").ends_with("homeboy"));
-    }
-
-    #[test]
-    fn test_runner_session_file_uses_id_filename() {
-        let path = runner_session_file("lab-box").expect("runner_session_file resolves");
-        assert_eq!(
-            path.file_name().and_then(|s| s.to_str()),
-            Some("lab-box.json")
-        );
-        assert_eq!(
-            path.parent()
-                .and_then(|p| p.file_name())
-                .and_then(|s| s.to_str()),
-            Some("runner-sessions")
-        );
     }
 }
