+++ crates/homeboy-paths/src/rigs.rs
